Testing implemented

+++ src/controller/department.controller.ts
@@ -17,6 +17,7 @@ import { RequestWithUser } from "../utils/requestwithuser";
 import EntityNotFoundException from "../exceptions/entitiynotfound.exception";
 import { Role } from "../utils/role.enums";
 import { ErrorCodes } from "../utils/error.codes";
+import { error } from "console";
 
 class DepartmentController {
   //   private employeeService: EmployeeService;
@@ -26,40 +27,58 @@ class DepartmentController {
     // this.employeeService = new EmployeeService();
     this.router = express.Router();
 
-    this.router.get("/", this.getAllDepartments);
-    this.router.get("/:id", this.getDepartmentById);
-    this.router.put("/:id", this.updateDepartment);
+    this.router.get("/", authorize, this.getAllDepartments);
+    this.router.get("/:id", authorize, this.getDepartmentById);
+    this.router.put("/:id", authorize, this.updateDepartment);
     this.router.post("/", authorize, this.createDepartment);
     this.router.delete("/:id", authorize, this.removeDepartment);
   }
 
-
   public getAllDepartments = async (
-    req: express.Request,
-    res: express.Response
-  ) => {
-    const department = await this.departmentService.getAllDepartments();
-    res.status(200).send(department);
-  };
-
-  public getDepartmentById = async (
-    req: express.Request,
+    req: RequestWithUser,
     res: express.Response,
     next: express.NextFunction
   ) => {
     try {
-      const departmentId = Number(req.params.id);
-      const department = await this.departmentService.getDepartmentById(
-        departmentId
-      );
-      if (!department) {
-        const error = new HttpException(
-          404,
-          `No department found with the ID : ${req.params.id}`
-        );
-        throw error;
+      const role = req.role;
+
+      if (role != Role.HR && role != Role.CEO) {
+        console.log(req.name);
+        throw new EntityNotFoundException(ErrorCodes.UNAUTHORIZED);
+      } else {
+        const department = await this.departmentService.getAllDepartments();
+        res.status(200).send(department);
       }
-      res.status(200).send(department);
+    } catch (err) {
+      next(err);
+    }
+  };
+
+  public getDepartmentById = async (
+    req: RequestWithUser,
+    res: express.Response,
+    next: express.NextFunction
+  ) => {
+    try {
+      const role = req.role;
+
+      if (role != Role.HR && role != Role.CEO) {
+        console.log(req.name);
+        throw new EntityNotFoundException(ErrorCodes.UNAUTHORIZED);
+      } else {
+        const departmentId = Number(req.params.id);
+        const department = await this.departmentService.getDepartmentById(
+          departmentId
+        );
+        if (!department) {
+          const error = new HttpException(
+            404,
+            `No department found with the ID : ${req.params.id}`
+          );
+          throw error;
+        }
+        res.status(200).send(department);
+      }
     } catch (err) {
       next(err);
     }
@@ -115,11 +134,18 @@ class DepartmentController {
             `No department found with the ID : ${req.params.id}`
           );
           throw error;
+        } else if (!department.employee) {
+          const departments = await this.departmentService.removeDepartment(
+            departmentId
+          );
+          res.status(204).send(departments);
+        } else {
+          const error = new HttpException(
+            406,
+            `Cannot delete department as employees exist`
+          );
+          throw error;
         }
-        const departments = await this.departmentService.removeDepartment(
-          departmentId
-        );
-        res.status(204).send(departments);
       }
     } catch (err) {
       next(err);
@@ -127,22 +153,29 @@ class DepartmentController {
   };
 
   public updateDepartment = async (
-    req: express.Request,
+    req: RequestWithUser,
     res: express.Response,
     next: express.NextFunction
   ) => {
     try {
-      const departmentDto = plainToInstance(CreateDepartmentDto, req.body);
-      const errors = await validate(departmentDto);
-      if (errors.length) {
-        console.log(errors);
-        throw new ValidationException(400, "Validation Failed", errors);
+      const role = req.role;
+
+      if (role != Role.HR && role != Role.CEO) {
+        console.log(req.name);
+        throw new EntityNotFoundException(ErrorCodes.UNAUTHORIZED);
+      } else {
+        const departmentDto = plainToInstance(CreateDepartmentDto, req.body);
+        const errors = await validate(departmentDto);
+        if (errors.length) {
+          console.log(errors);
+          throw new ValidationException(400, "Validation Failed", errors);
+        }
+        const departments = await this.departmentService.updateDepartment(
+          Number(req.params.id),
+          departmentDto.name
+        );
+        res.status(200).send(departments);
       }
-      const departments = await this.departmentService.updateDepartment(
-        Number(req.params.id),
-        departmentDto.name
-      );
-      res.status(200).send(departments);
     } catch (err) {
       next(err);
     }
